Migration to replace GitPreparer by GitRepositoryFactory
