+++ src/GitTools.Core/GitTools.Core.Shared/Git/GitRepositoryFactory.cs
@@ -1,4 +1,4 @@
-namespace GitTools.Git
+﻿namespace GitTools.Git
 {
     using System;
     using System.IO;
@@ -6,51 +6,51 @@ namespace GitTools.Git
     using LibGit2Sharp;
     using Logging;
 
-    public class GitPreparer
+    public static class GitRepositoryFactory
     {
         private static readonly ILog Log = LogProvider.GetCurrentClassLogger();
 
-        string targetUrl;
-        string dynamicRepositoryLocation;
-        AuthenticationInfo authentication;
-        bool noFetch;
-        string targetPath;
-
-        public GitPreparer(string targetUrl, string dynamicRepositoryLocation, AuthenticationInfo authentication, bool noFetch, string targetPath)
+        /// <summary>
+        /// Creates the repository based on the repository info. If the <see cref="RepositoryInfo.Directory"/> points 
+        /// to a valid directory, it will be used as the source for the git repository. Otherwise this method will create
+        /// a dynamic repository based on the url and authentication info.
+        /// </summary>
+        /// <param name="repositoryInfo">The repository information.</param>
+        /// <param name="noFetch">If set to <c>true</c>, don't fetch anything.</param>
+        /// <returns>The git repository.</returns>
+        public static GitRepository CreateRepository(RepositoryInfo repositoryInfo, bool noFetch = false)
         {
-            this.targetUrl = targetUrl;
-            this.dynamicRepositoryLocation = dynamicRepositoryLocation;
-            this.authentication = authentication ?? new AuthenticationInfo();
-            this.noFetch = noFetch;
-            this.targetPath = targetPath;
-        }
+            bool isDynamicRepository = false;
+            string repositoryDirectory = null;
 
-        public bool IsDynamicGitRepository
-        {
-            get { return !string.IsNullOrWhiteSpace(DynamicGitRepositoryPath); }
-        }
-
-        public string DynamicGitRepositoryPath { get; private set; }
-
-        public void Initialise(bool normaliseGitDirectory, string currentBranch)
-        {
-            if (string.IsNullOrWhiteSpace(targetUrl))
+            // TODO: find a better way to check for existing repositories
+            if (!string.IsNullOrWhiteSpace(repositoryInfo.Directory))
             {
-                if (normaliseGitDirectory)
-                {
-                    GitHelper.NormalizeGitDirectory(GetDotGitDirectory(), authentication, noFetch, currentBranch);
-                }
-                return;
+                repositoryDirectory = repositoryInfo.Directory;
+            }
+            else
+            {
+                isDynamicRepository = true;
+
+                var tempRepositoryPath = CalculateTemporaryRepositoryPath(repositoryInfo.Url, repositoryDirectory);
+                repositoryDirectory = CreateDynamicRepository(tempRepositoryPath, repositoryInfo.Authentication,
+                    repositoryInfo.Url, repositoryInfo.Branch, noFetch);
             }
 
-            var tempRepositoryPath = CalculateTemporaryRepositoryPath(targetUrl, dynamicRepositoryLocation);
+            // TODO: Should we do something with fetch for existing repositoriess?
 
-            DynamicGitRepositoryPath = CreateDynamicRepository(tempRepositoryPath, authentication, targetUrl, currentBranch, noFetch);
+            var repository = new Repository(repositoryDirectory);
+            return new GitRepository(repository, isDynamicRepository);
         }
 
         static string CalculateTemporaryRepositoryPath(string targetUrl, string dynamicRepositoryLocation)
         {
-            var userTemp = dynamicRepositoryLocation ?? Path.GetTempPath();
+            var userTemp = dynamicRepositoryLocation;
+            if (string.IsNullOrWhiteSpace(userTemp))
+            {
+                userTemp = Path.GetTempPath();
+            }
+
             var repositoryName = targetUrl.Split('/', '\\').Last().Replace(".git", string.Empty);
             var possiblePath = Path.Combine(userTemp, repositoryName);
 
@@ -88,31 +88,11 @@ namespace GitTools.Git
             }
         }
 
-        public string GetDotGitDirectory()
-        {
-            if (IsDynamicGitRepository)
-            {
-                return DynamicGitRepositoryPath;
-            }
-
-            return GitDirFinder.TreeWalkForDotGitDir(targetPath);
-        }
-
-        public string GetProjectRootDirectory()
-        {
-            if (IsDynamicGitRepository)
-            {
-                return targetPath;
-            }
-
-            return Directory.GetParent(GitDirFinder.TreeWalkForDotGitDir(targetPath)).FullName;
-        }
-
         static string CreateDynamicRepository(string targetPath, AuthenticationInfo authentication, string repositoryUrl, string targetBranch, bool noFetch)
         {
             if (string.IsNullOrWhiteSpace(targetBranch))
             {
-                throw new Exception("Dynamic Git repositories must have a target branch (/b)");
+                throw new GitToolsException("Dynamic Git repositories must have a target branch (/b)");
             }
 
             Log.Info(string.Format("Creating dynamic repository at '{0}'", targetPath));
@@ -164,18 +144,18 @@ namespace GitTools.Git
                 var message = ex.Message;
                 if (message.Contains("401"))
                 {
-                    throw new Exception("Unauthorised: Incorrect username/password");
+                    throw new GitToolsException("Unauthorised: Incorrect username/password");
                 }
                 if (message.Contains("403"))
                 {
-                    throw new Exception("Forbidden: Possbily Incorrect username/password");
+                    throw new GitToolsException("Forbidden: Possbily Incorrect username/password");
                 }
                 if (message.Contains("404"))
                 {
-                    throw new Exception("Not found: The repository was not found");
+                    throw new GitToolsException("Not found: The repository was not found");
                 }
-                
-                throw new Exception("There was an unknown problem with the Git repository you provided");
+
+                throw new GitToolsException("There was an unknown problem with the Git repository you provided");
             }
         }
     }
