+++ src/GitTools.Core/GitTools.Core.Shared/Git/RepositoryInfo.cs
@@ -1,0 +1,18 @@
+﻿namespace GitTools.Git
+{
+    public class RepositoryInfo
+    {
+        public RepositoryInfo()
+        {
+            Authentication = new AuthenticationInfo();
+        }
+
+        public AuthenticationInfo Authentication { get; set; }
+
+        public string Directory { get; set; }
+
+        public string Branch { get; set; }
+
+        public string Url { get; set; }
+    }
+}
